Move azure_storage_utils
