+++ detection.Dockerfile
@@ -28,7 +28,8 @@ RUN python -m pip install -e detectron2
 
 WORKDIR /app
 COPY configs/* /app/configs/
-COPY inference.py evaluation.py utils.py azure_storage_utils.py /app/
+COPY inference.py evaluation.py utils.py /app/
+COPY utils/azure_storage.py /app/utils
 # copy weights
 # COPY model_final.pth /app/
