+++ submit_to_sia.py
@@ -11,7 +11,7 @@ import argparse
 import pandas.io.sql as sqlio
 
 import upload_to_postgres
-from azure_storage_utils import BaseAzureClient, StorageAzureClient
+from utils.azure_storage import BaseAzureClient, StorageAzureClient
 
 BASE_URL = "https://acc.api.meldingen.amsterdam.nl/signals/v1/private/signals"
 API_MAX_UPLOAD_SIZE = 20 * 1024 * 1024  # 20MB = 20*1024*1024
