+++ upload_to_postgres.py
@@ -14,7 +14,7 @@ from psycopg2._psycopg import cursor  # pylint: disable-msg=E0611
 from psycopg2.errors import ConnectionException  # pylint: disable-msg=E0611
 from psycopg2.extras import execute_values
 
-from azure_storage_utils import BaseAzureClient, StorageAzureClient
+from utils.azure_storage import BaseAzureClient, StorageAzureClient
 
 azClient = BaseAzureClient()
 USERNAME = azClient.get_secret_value("postgresUsername")
